# backend/nosd/openapi.yaml
openapi: 3.0.3
info:
  title: NithronOS API
  version: 0.1.0
  description: API for nOS backend
servers:
  - url: http://127.0.0.1:9000
paths:
...
  /api/pools/plan-create:
    post:
      summary: Plan Btrfs pool creation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolCreateReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
...
  /api/pools/import:
    post:
      summary: Import (mount) a pool by device or UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_or_uuid: { type: string }
              required: [device_or_uuid]
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pools/create:
    post:
      summary: Create Btrfs pool
      parameters:
        - in: header
          name: Confirm
          schema: { type: string }
          required: true
          description: must be "yes" for destructive operations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolCreateReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Precondition Required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pools/{id}/snapshots:
    get:
      summary: List snapshots
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Snapshot'
    post:
      summary: Create snapshot
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subvol: { type: string }
                name: { type: string }
              required: [subvol]
      responses:
        '200':
          description: OK
  /api/shares:
    get:
      summary: List shares
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Share'
    post:
      summary: Create share
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Share'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Share'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/shares/{id}:
    delete:
      summary: Delete share
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/smb/users:
    get:
      summary: List SMB users (proxy)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringArray'
    post:
      summary: Create SMB user (proxy)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmbUserCreate'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/updates/apply:
    post:
      summary: Apply updates, optionally snapshotting targets before
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                packages:
                  type: array
                  items: { type: string }
                snapshot:
                  type: boolean
                confirm:
                  type: string
              required: [confirm]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  tx_id: { type: string }
                  snapshots_count: { type: integer }
                  updates_count: { type: integer }
                required: [ok, tx_id]
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/updates/rollback:
    post:
      summary: Roll back a prior updates transaction using its snapshots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tx_id: { type: string }
                confirm: { type: string }
              required: [tx_id, confirm]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                required: [ok]
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            application/octet-stream:
              schema:
                type: string
                format: binary
...
  /api/firewall/plan:
    post:
      summary: Plan firewall rules for a mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode: { type: string, enum: [lan-only, vpn-only, tunnel, direct] }
              required: [mode]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules: { type: string }
                required: [rules]
  /api/firewall/apply:
    post:
      summary: Apply firewall rules for a mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode: { type: string, enum: [lan-only, vpn-only, tunnel, direct] }
                twoFactorToken: { type: string, nullable: true }
              required: [mode]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                required: [ok]
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    StringArray:
      type: array
      items: { type: string }
    SmbUserCreate:
      type: object
      properties:
        username: { type: string }
        password: { type: string, nullable: true }
      required: [username]
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Snapshot:
      type: object
      properties:
        path: { type: string }
        name: { type: string }
        readonly: { type: boolean }
    PoolCreateReq:
      type: object
      properties:
        label: { type: string }
        devices:
          type: array
          items: { type: string }
        raid: { type: string, enum: [single, raid1, raid10] }
      required: [label, devices, raid]
    Share:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [smb, nfs]
        path:
          type: string
        name:
          type: string
        ro:
          type: boolean
        users:
          type: array
          items: { type: string }
      required: [type, path, name]
...